+++ src/utils/LanguageCheck.js
@@ -1,15 +1,17 @@
 import React, { useLayoutEffect } from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { languageChanged } from 'store/uiSlice';
 
 const LanguageCheck = ({ children }) => {
+    const dispatch = useDispatch();
+
     const { appSettings } = useSelector(({ ui }) => ui);
 
     useLayoutEffect(() => {
         if (appSettings?.isLanguageCheckTriggered) {
             const lang = navigator?.language?.split('-')[0];
-            if (appSettings?.supportedLanguages.includes(lang)) languageChanged(lang);
-            else languageChanged('default');
+            if (appSettings?.supportedLanguages.includes(lang)) dispatch(languageChanged(lang));
+            else dispatch(languageChanged('default'));
         }
     }, [appSettings?.isLanguageCheckTriggered, appSettings?.supportedLanguages]);